Refactored instances start and stop functionality
